Revert "LF-301: dmaengine: imx-sdma: Add once more loading firmware"
This was added for cases where the booting process could take over a minute,
such as booting from NFS on an i.MX6SLL. In our case, it was causing forced
reboot operations to sometimes delay for an additional minute in our recovery
ramdisk, since it doesn't include the SDMA fw.

Remove the retry to avoid this scenario, because all of our i.MX8 platforms
take less than a minute to boot, and the ccimx6ul has the SDMA fw integrated
into its kernel.

This reverts commit 5ebe728cc7de44282212956b9187c8f1155fb99a.

https://jira.digi.com/browse/DEL-7152

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>