ARM: dts: ccimx8x-sbc-pro: fix goodix touch pinctrl configuration
The pinctrl configuration of the touch interrupt line belongs to the
touch driver and not to the lvds interface. Otherwise the driver may
access the gpios (reset and irq) before the pinctrl has been configured,
failing in that case.
This could result in a initialization failure of the touch controller,
failing to set the i2c address of the chip.

This problem has been observed on the ConnectCore 8X SBCPro boards v4
and v5 where the irq line is being pull-up by a resistor and the touch
driver has to pull it down during the driver probe phase.

https://jira.digi.com/browse/CC8X-271

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>
(cherry picked from commit 290d39a991460f750be7b20104499a4f441fb716)
Signed-off-by: Javier Viguera <javier.viguera@digi.com>