drivers: goodix: add 60 ms delay in resume function
When the touch controller is power cycled during the suspend mode, sometimes
it fails during its reinitialization when resuming the system from suspend.
This seems to be caused when the int gpio, which is used to set the address of
the device, is toggled too early after applying power. A small delay before
toggling the int gpio helps to overcome this problem.

https://jira.digi.com/browse/CC8X-271

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>