dt-bindings: dma: Document STM32 DMA3 controller bindings
The STM32 DMA3 is a Direct Memory Access controller with different features
depending on its hardware configuration.
The channels have not the same capabilities, some have a larger FIFO, so
their performance is higher.
This patch describes STM32 DMA3 bindings, used to select a channel that
fits customer requirements, and to pre-configure the transfer depending on
the customer needs.

Signed-off-by: Amelie Delaunay <amelie.delaunay@foss.st.com>
Change-Id: I475a24921b51823fd03b0de6f47d8e7699f07427