ARM64: dts: ccimx8x-sbc-pro: fix active level of goodix touch irq line
Set the proper active level on the touch interrupt gpio line, if not, when set
it is automatically inverted by the gpio subsystem what can cause
initialization problems of the touch controller.

Note that the irq line is used by the driver to initialize the i2c address of
the touch controller.

https://jira.digi.com/browse/DEL-7241

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>