ccimx8: dts: keypad: split the keypad overlay into two
Due to is impossible to have the same overlay to testing the keypad
we have decided to split it into two differents overlays, one for
the ccimx8mn and another one for the ccimx8x.

In the ccimx8mn we need to disable the XBee cellular in order to
have some free pins for the keypad.

Signed-off-by: Francisco Gil Martinez <francisco.gilmartinez@digi.com>
(cherry picked from commit 11f53b1cea6e5a06dd2812945387fa983fa9b4af)