goodix: leave interrupt GPIO configured as input when going to suspend
Using specific hardware combinations of the ccimx8x SBC Pro and the video
adapter board causes the touch controller to never awaken from suspend, due to
the IRQ line not being toggled with enough voltage. To fix this, set the GPIO
as input when going to suspend to let the 3.3V pull-up drive the line high
in case the signal provided by the CPU isn't strong enough.

https://jira.digi.com/browse/DEL-7276

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>