ARM64: dts: ccimx8mnano-dvk: fix RS485 RE#/DE line for DVKv3
On v1 and v2 of the DVK, the pin to control RS485 RE#/DE line was GPIO1_IO3
configured as gpio.
On following designs, the pin was changed to GPIO5_IO08 which has dedicated
UART3_CTS_B functionality.

Note that this commit does not make use of the pin dedicated functionality and
keeps on using it as gpio.

https://jira.digi.com/browse/DEL-6945

Signed-off-by: Hector Bujanda <hector.bujanda@digi.com>
(cherry picked from commit 7bfb04653e3530fcfc6c9d5000178dc2dfe75012)