ARM64: dts: ccimx8mn: separate HPD gpio IOMUX pinctrl
The GPIO used for HDMI hotplug detect on DVKv3 was used for
MCA interrupt on SOMv1. This produced a clash between the
board v1v2 overlay and the SOMv1 overlay which resulted on
HDMI not working on the boards with this combination on SOM
and DVK.

Take the HPD pinctrl to another group and make the overlay
configure the bridge only for the IOMUX of other pads.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
(cherry picked from commit 8086dc5f8e298a5df8fde6f925f3d7afda9f7bef)