serial: stm32: prevent TDR register overwrite when sending x_char
When sending x_char in stm32_usart_transmit_chars(), driver has a chance
to overwrite the value of TDR register by the value of x_char. If this
happens, the previous value that was present in TDR register will not be
sent through uart.

This code checks if the previous value in TDR register is sent before
writing the x_char value into register.

Fixes: 48a6092fb41f ("serial: stm32-usart: Add STM32 USART Driver")
Signed-off-by: Valentin Caron <valentin.caron@foss.st.com>