ARM64: dts: ccimx8x: add missing properties to the CMA reserved memory node
Otherwise, the kernel won't parse the digi,size-table property and the same
size of 320 MB will be used on all variants of the SOM, regardless of RAM size.
This node used to be present in the SOC .dtsi files, but it was moved to the
NXP MEK .dtsi, so copy all of its properties into the SOM .dtsi so the kernel
can parse through it correctly.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>