ARM64: dts: ccimx8x: add missing regulator and lines for usb modems
Before creating the overlay for the PCIe we have all the things
defined in the ccimx8x-sbc-pro.dts so everything was working fine.

After the migration to the overlay system, we miss the reg_3v3_pcie
needed for the modems power rail.

The reset and disable GPIOs are mandatory for TELIT modems without them
the modems are not recognized.

Signed-off-by: Francisco Gil Martinez <francisco.gilmartinez@digi.com>