ARM64: dts: ccimx8x-sbc-pro: assert MCA_IO9 high on SBCv4
MCA_IO9, which was free on SBC Pro v1-v3, is used on v4 for
cutting the power to the PMIC so it needs to be reserved and
asserted high to prevent the system from powering off.
This is done by a dummy always-on regulator.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-7338
(cherry picked from commit 2e23cc2e9681cf28717c2f3ee03b301e3417a696)
Signed-off-by: Hector Palacios <hector.palacios@digi.com>