imx-ocotp: initialize driver as early as possible
On ccimx6/ccimx6+/ccimx6ul platforms, root filesystem encryption uses the
device's HWID as a key modifier, which is accessed via the OTP driver. On the
ccimx6ul, filesystem encryption is done at the mtd driver level, and the OTP
bits must be accessible by the time the mtd partitions are parsed to be able to
use the HWID key modifier.

To avoid situations where encrypted mtd partitions are parsed before the OTP
driver is ready, which can cause unexpected behaviour, make sure the OTP driver
is initialized as early as possible.

https://jira.digi.com/browse/DEL-7361

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>