ARM64: dts: ccimx8x-sbc-pro/ccimx8mn-dvk: use correct ov5640 MIPI camera driver
Even though the ov5640_mipi_v3 driver works in most cases, it causes unexpected
behaviour when trying to record on the 8X in HD. This is because said driver
comes from the v4.14 BSP and isn't fully in sync with the video subsystem in
v5.4.

Rework the ov5640_mipi nodes in the ccimx8x and ccimx8mn boards to match what
NXP is doing in their reference designs and make use of the standard ov5640
driver instead.

https://jira.digi.com/browse/CC8X-282

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>