dmaengine: stm32-dma: update stm32_dma_set_xfer_param prototype
buf_addr parameter of stm32_dma_set_xfer_param function is a dma_addr_t but
the final use of this parameter (in stm32_dma_get_max_width()) is to use it
as a u64 for do_div().
So, move buf_addr parameter from dma_addr_t to u64, add the explicit cast
to u64 when calling stm32_dma_set_xfer_param and remove the cast to u64
when calling stm32_dma_get_max_width because buf_addr is already a u64.

Change-Id: I40f2fed694b4152f0bc8cfcfe824584c7d02d3d4
Signed-off-by: Amelie Delaunay <amelie.delaunay@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/189891
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: Fabien DESSENNE <fabien.dessenne@st.com>