Revert "media: ov5640: Add 60 fps support"
This reverts commit e823fb165b763bf0e42bceb9ab45cb9f7589d871.

This commit introduced VGA@60fps support but leads to regressions
and unstability on application side due to the fact that resolution
change depends on supported framerate for this resolution, so if unlikely
VGA@60fps has been selected at a time, resolution change can occur only if
VGA@15fps or VGA@30fps is chosen first. This is not intuituve so better to
drop VGA@60fps for the time being.

NOT_UPSTREAMABLE

Signed-off-by: Hugues Fruchet <hugues.fruchet@st.com>
Change-Id: If440c71474f9562aa2ea9ef0555da537401b2be4