dmaengine: stm32-mdma: check the channel availability (secure or not)
On STM32MP15, STM32_MDMA_CCR bit[8] is not documented but exists.
On STM32MP13, STM32_MDMA_CCR bit[8] is documented.
So, no need to add this bit support only for STM32MP13.

STM32_MDMA_CCR bit[8] is used to enable Secure Mode (SM). If this bit is
set, it means that all the channel registers are write-protected. So the
channel is not available for Linux use.

Add stm32_mdma_filter_fn() callback filter and give it to
__dma_request_chan (instead of dma_get_any_slave_channel()), to exclude the
channel if it is marked Secure.

Change-Id: Ia245f0faebc6a4f3e82e9bb5e0b46f8c44f764d7
Signed-off-by: Amelie Delaunay <amelie.delaunay@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/190166
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Fabien DESSENNE <fabien.dessenne@st.com>