ARM: dts: ccimx6ulsbc: add display 5v regulator and remove reset from touch
Commit 18828fe2986924147db391d569d17229360a5f30 added IOEXP_29 as the
reset-gpios property to the touch controller as a work-around to have
a stable power sequence and i2c address latching of the touch controller.

IOEXP_29 is not routed however to the reset line of the touch. This gpio
is actually controlling the 5V power supply to the LCD/LVDS connector.
On Video Adapter Board (VAB) v2, the touch was powered to this line (and
the reset line shortcircuited to it), so the effect of having this property
was the power-cycling of the touch controller (and any 5V circuitry of the
display, including the backlight).

On VABv3, the touch is powered from 3V3, so keeping the reset line on the
DT simply toggles the 5V (backlight glitch) but doesn't have any effect
over the touch controller.

IOEXP_29 however needs to be enabled, not as the reset line of the touch,
but as the main supply of the display. Having an LVDS bridge however, that
also needs to be powered (by a different line) forces us to declare two
supplies:
- the 5V (new regulator) controlled by IOEXP_29 is the main supply.
- the LDO4 (that powers the bridge) is the auxiliar supply.

https://jira.digi.com/browse/DEL-7241

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
(cherry picked from commit 880e6fe723356cf0b9b9a662acb0eea43ca70289)