ARM: dts: ccimx6ulsbc: duplicate touch entry at 0x5d
The Video Adapter Board (VAB) doesn't have control over the RESET line of
the touch (short-circuited to the touch power line). This line is used in
combination with the INT line to determine the I2C address of the touch
controller.
In certain scenarios (such as when coming from a soft-reset), the touch
may initialize on address 0x5d (instead of 0x14).

Add a second touch entry at 0x5d so that the touch is always detected in
either of the two possible addresses.
Inevitably, the kernel will report that it couldn't find one of them.

https://jira.digi.com/browse/DEL-7241

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>
(cherry picked from commit 69f5fe438390b6cea0499de7dfcdfbde600a73ab)