ARM64: dts: ccimx8m: duplicate touch entry at 0x5d
The Video Adapter Board (VAB) doesn't have control over the RESET line of
the touch (short-circuited to the touch power line). This line is used in
combination with the INT line to determine the I2C address of the touch
controller.
In certain scenarios (such as when coming from a soft-reset), the touch
may initialize on address 0x5d (instead of 0x14).

Add a second touch entry at 0x5d so that the touch is always detected in
either of the two possible addresses.
Inevitably, the kernel will report that it couldn't find one of them.

https://jira.digi.com/browse/DEL-7241

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>
(cherry picked from commit 47ef853e9c6c2197b52ecd4e60f7eabeb0ab00bd)