non-upstream: Revert "brcm80211: select WANT_DEV_COREDUMP conditionally for brcmfmac"
This reverts commit 885a93cf31106808cd3ded7c5e01e35fa495fdc5.

To fix DEV_COREDUMP function to be disabled in compat/Kconfig
of backports. The root cause is because the parsing rule of
lib/kconfig.py in backports will not consider the "select <sym>
if xxx" as a sel_line. In order to aviod affecting other driver's
configs, so we do not modify the sel_line logic in backports.

Signed-off-by: Ting-Ying Li <tingying.li@infineon.com>
Signed-off-by: Chung-Hsien Hsu <chung-hsien.hsu@infineon.com>
Signed-off-by: Chi-hsien Lin <chi-hsien.lin@infineon.com>