touch: fusion: defer probe if the ID can't be read
The Fusion touch controller doesn't have a regulator, but the kernel may
try to access the touch before enabling the regulator of the display
(which powers the touch as well).
For this reason, we want to defer probe a number of times if during the
probe of the driver we fail to identify the touch.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-7464