touch: fusion: defer probe if the ID can't be read
The Fusion touch controller doesn't have a regulator, but the kernel may
try to access the touch before enabling the regulator of the display
(which powers the touch as well).
For this reason, we want to defer probe a number of times if during the
probe of the driver we fail to identify the touch.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-7464
(cherry picked from commit 15ce814416e7d3464e275cd2f4bb12a65d268163)
(cherry picked from commit 536ffa101b7c5b3f01cec0bcc2cda9fe5a36ce25)