mca-wdt: fix mca_wdt_stop() so it retuns error if watchdog configured as no-way-out
The mca_wdt_stop() function was sending the command to disable the watchdog and
returning 0 without reading it back to check if it applied.
If the watchdog was previously configured as no-way-out, the MCA won't disable
the watchdog and won't clear corresponding flag.
The change in this commit reads back the watchdog control register and returns
an error if the function didn't manage to stop the watchdog.

Note that, if the kernel was built with CONFIG_WATCHDOG_NOWAYOUT=y, Linux higher
layers will report an error without ever calling mca_wdt_stop() function.

https://jira.digi.com/browse/DUB-815

Signed-off-by: Hector Bujanda <hector.bujanda@digi.com>
(cherry picked from commit 121b5e174b6c8b6da173cb712f52a4c20341cd4d)
Signed-off-by: Javier Viguera <javier.viguera@digi.com>