ccimx8x-sbc-pro.dtsi: fix wifi pci support
commit a3e1d1a7f5fc adds support to validate ancestors DT bindings
like clocks. On commit 5c5de3e6c1f0 we removed the associated GPIOs for
hsio_refb_clk to be platform specific (that is ok). But now, with this
new kernel check, it fails. We do not need to enable the hsio_refb_clk by
using a GPIO but we need to use the clock to have the PCI working. This
commit fixes it.

https://jira.digi.com/browse/ADK4A-1803

Signed-off-by: Javier Viguera <javier.viguera@digi.com>