drivers: of: reset driver state to DEY version 5.4.47
Reset the OF driver to the status it had in DEY kernel 5.4.47 commit
1e23679a9d3a30f8e4b79a8bc41f51a82b026198

This driver was giving several issues with device trees and the new look for ancestors
functionality. It was causing some GPIOs to be incorrectly held by other drivers not
defining them, for example the PCIE or MIPI camera interfaces.

Signed-off-by: David Escalona <david.escalona@digi.com>