drivers: gpio-mca: fix interrupts on multiple instances
When multiple instances of the gpio-mca are used, for instance, when the
io-expander of the cc6ulsbc is present, the interrupt support will not
work on the 2nd and later instances.

The reason is that the driver is sharing/reusing the same irq_chip data
structure, which is not allowed in newer kernels.

The fix provides an irq_chip structure per instance.

https://onedigi.atlassian.net/browse/DEL-7541

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>
(cherry picked from commit ea5a23418a8dba030d305e15604ab511c2f81c5f)