ARM: dts: stm32: manage v3v3_sw hub supply in on stm32mp135f-dk
USB hub of stm32mp135f-dk can't be managed because I2C interface is not
wired.
USB hub reset and VBUS_DET is wired on v3v3_sw.
The point is, when suspending the platform, v3v3_sw is powered off and the
hub DP/DM lines trigger the wakeup condition when a USB wakeup source is
enabled, so the platform automatically wakes up when v3v3_sw is powered
off.

The workaround here, is to manage v3v3_sw supply in usbphyc_port0 node
(on which the hub is wired). In usbphyc driver, "vbus-supply" is enabled
on phy_power_on and disabled on phy_power_off.
When a device is wakeup source, phy_power_off is not called, then
"vbus-supply" remains enabled.
Thus, the platform can suspend, and USB wakeup source can wake up the
platform.

Change-Id: Ie023fdf66dfd3d14d9546c507674622c0c3c2a7f
Signed-off-by: Amelie Delaunay <amelie.delaunay@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/229465
Tested-by: Eric FOURMONT <eric.fourmont-ext@st.com>
Reviewed-by: Eric FOURMONT <eric.fourmont-ext@st.com>