ARM64: dts: ccimx8m: increase critical trip point to 100C
The SOC dtsi establishes the critical trip point at 95C.
All Digi SOMs are industrial grade (max 105C) so 100C is enough
limit for a critical shut down.

The commit adds also the passive trip point and other parameters
for increased readability, though the only overriden property with
respect to the SOC dtsi is the temperature limit of the critical
trip point.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>