ARM: dts: stm32: manage v3v3 hub supply on stm32mp15xx-evx
USB hub of stm32mp15xx-evx can't be managed because I2C interface is not
wired.

USB hub reset and VBUS_DET is wired on v3v3.
The point is, when suspending the platform, v3v3 is powered off and
devices connected to the hub may not be suspended yet.

The workaround here, is to manage v3v3 supply in usbphyc_port0 node
(on which the hub is wired). In usbphyc driver, "vbus-supply" is enabled
on phy_power_on and disabled on phy_power_off.

When suspending the platform, phy_power_off is called after suspending all
the connected devices if any of them has wakeup enabled.

Change-Id: I5fcc0b9ec1b6bd5e2f7c5b4fde9e92c893d41434
Signed-off-by: Amelie Delaunay <amelie.delaunay@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/229464
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Eric FOURMONT <eric.fourmont-ext@st.com>
Tested-by: Eric FOURMONT <eric.fourmont-ext@st.com>