wireless: ath6kl: update wmi_connect_event_common
In hostapd mode the current ath6kl_wmi_connect_event function
in kernel 5.4 is setting the same AID to all the stations (5) connected
to the AP, this is causing that the second station connected to the AP
is the only one accessible from the target although the first station
is still connected to the AP.

This is caused by a wrong order in the parameters of the
wmi_connect_event_common struct.

If the parameters are set like the upstream driver the issue is
reproducible, after the change in the structure, the issue disappear.

This commits reverts the one done in:
31ba6a088ed1739abe0ac5b305fc9b76bd70b39a

The driver starts diverging from upstream in this part since:
688e14efd7ad13ff233221b64d16db54dfd1662a

With this commit the driver is set in the same way that it was
in dey-2.6.

https://onedigi.atlassian.net/browse/DEL-7621

Signed-off-by: Francisco Gil <francisco.gilmartinez@digi.com>
(cherry picked from commit f894e2258ffdbaa6ba35d34f71abddbca5a45819)