drivers: lt8912: omit drm call when dev is uninitialized
Sometimes there is a race condition on the CCIMX8M Mini when both an HDMI
display and the MIPI camera are connected that can trigger the display
Hot-Plug detection interrupt (handler 'lt8912_hpd_threaded_handler')
before the display has been properly initialized in 'lt8912_bridge_attach'.
This can cause a Kernel error and stop the device initialization.

Omit the drm call in those cases to proceed with initialization as usual.

https://onedigi.atlassian.net/browse/DEL-7573

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>