imx_thermal: don't set PANIC_ALARM_VALUE
On i.MX6UL, register TEMPMON_TEMPSENSE2 has a PANIC_ALARM_VALUE field used to
configure a 'Panic threshold' which, when surpassed, forces a chip reset.

This threshold was configured with the same value as the 'Critical trip point'
which, when surpassed, triggers a controlled shutdown of the system.

On the CC6UL platform both conditions happen at the same time, so whenever
the chip temperature goes over the 'Critical trip point' the controlled
shutdown is initiated but abruptly interrupted by a hardware reset.

Avoid setting the PANIC_ALARM_VALUE threshold in order to exclusively
trigger the software-controlled shutdown, in a manner similar to the rest
of the platforms in the ConnectCore family.

https://onedigi.atlassian.net/browse/DEL-7630

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>
(cherry picked from commit 9e5a0b6ee1090a50b8e26f9b7de4aaff2e31be83)
Signed-off-by: Javier Viguera <javier.viguera@digi.com>