arm64: dts: ccimx8x-sbc-pro: change native CS (SPI) to a GPIO
When you try to send some data through the SPI interface, for
each byte transmited, the CS line is toggling it state.

With some devices the CS line should maintain the level during
the transmission (when sending more than a byte),
for example in the EEPROM SPI example provided in the
libdigiapix.

Using the CS line as GPIO solves the issue.

NXP is using the same approach in their dts:
imx8qxp-lpddr4-val-lpspi.dts

https://onedigi.atlassian.net/browse/DEL-7590

Signed-off-by: Francisco Gil <francisco.gilmartinez@digi.com>