brcmfmac: use request_firmware_direct for loading board-specific nvram
Loading a board-specific nvram file is optional. If the file is not
present, we will get a warning "Direct firmware load for ... failed with
error ...", have 60 seconds delay when
CONFIG_FW_LOADER_USER_HELPER_FALLBACK is enabled, and fall back to a
common one. However, users may treat this as a type of side effect and
the file is required.

Since it is optional, use request_firmware_direct() and leave an info
message in place to avoid the delay and confusion.

Signed-off-by: Chung-Hsien Hsu <stanley.hsu@cypress.com>
Signed-off-by: Ting-Ying Li <tingying.li@infineon.com>
Signed-off-by: Chi-hsien Lin <chi-hsien.lin@infineon.com>