nvmem: stm32: add warning when upper OTPs are updated
As the upper OTPs are ECC protected, they support only one word programming.
For a second modification of the 32 bits word, these ECC become invalid
and this OTP will be no more accessible, the shadowed value is invalid.

This patch adds a warning to indicate an upper OTP update, because this
operation is dangerous as OTP is not locked by the driver after the first
update to avoid a second update.

Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Change-Id: If9d151089532b4602be30e7db73c980ad95cec8f