dmaengine: stm32-dma: fix chaining deadlock issue in case of MDMA threaded IRQ
For DMA/MDMA chaining mode, as soon as a transfer finishes, the MDMA GIC
interrupt is raised and then MDMA isr is called. Regarding MDMA status
register, the MDMA isr code schedules a tasklet to complete the DMA
transfer. At the execution of this tasklet (softirq),
stm32_dma_mdma_flush_remaining is called to check if there are some
remaining data in SRAM and if yes, starts a new transfer and waits
transfer end by spinning the CPU.
At the end of this transfer MDMA interrupt is raised and MDMA isr is then
called again to the "active wait" done in stm32_dma_mdma_flush_remaining.

When the MDMA isr is called under interrupt context (which is the case on
our mainline Kernel) everything is fine. We have:

mdma_isr
	-->tasklet_execution
		-->new transfer started-->active wait on transfer end
			mdma_isr-->complete the transfer

But we have a deadlock when MDMA isr is executed in a thread (which is the
case in Linux-RT kernel or if force_irqthreads is set). In this case the
MDMA isr AND the execution of the tasklet are done in the same thread. The
consequence is the CPU spinning in stm32_dma_mdma_flush_remaining will
"block" the MDMA isr thread and any new MDMA interrupt can't be serviced
and so stm32_dma_mdma_flush_remaining will wait up to a timeout.

This patch moves the tasklet code in a work in order to release the MDMA
ISR thread.

Fixes: 73c0a39c5d48 ("dmaengine: stm32-dma: Add DMA/MDMA chaining support")
Change-Id: I1c5d8bb6f277cf7983f163a4cae99b8c68fde2bd
Signed-off-by: Alexandre Torgue <alexandre.torgue@foss.st.com>
Signed-off-by: Amelie Delaunay <amelie.delaunay@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/241891
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Fabien DESSENNE <fabien.dessenne@foss.st.com>