nvmem: stm32-romem: detect op-tee presence for stm32mp15
On STM32MP15, the SMC backend is optional when OP-TEE is used;
the PTA BSEC should be used as it is done on STM32MP13 platform,
but the BSEC SMC can be also used: it is a legacy mode in OP-TEE,
not recommended but used in previous OP-TEE firmware.

The presence of OP-TEE is dynamically detected in STM32MP15x device tree
and the supported NVEM backend is dynamically detected:
- PTA with stm32_bsec_pta_find
- SMC with stm32_bsec_check

On STM32MP13x plaform, only the the PTA is supported (cfg->ta = true).

Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Change-Id: I9275f92e643dfb4c67f95e7c6749366d57664829
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/224607
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Lionel DEBIEVE <lionel.debieve@foss.st.com>