ARM: dts: stm32: add dfsdm support on stm32mp135f-dk
Add support of audio record through digital microphones
plugged on STM32MP135F Disco board expansion connector.
The DFSDM configuration assumes that two digital microphones
are plugged on DFSDM_DATIN1 and DFSDM_DATIN3 inputs and
are configured for right and left channels respectively.
In this example the microphones have also to be fed with
the clock from DFSDM_CKOUT pin.

Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/282198
Signed-off-by: Olivier Moysan <olivier.moysan@foss.st.com>
Change-Id: If56e47042e49c43ae4046cf76942956f62e01d70