pinctrl: stm32: fix optional IRQ support to gpios
To act as an interrupt controller, a gpio bank rely on the interrupt-parent
of the pin controller. When this optional interrupt-parent misses, do not
create any IRQ domain.

This fixes a "NULL pointer in stm32_gpio_domain_alloc()" kernel crash when
the interrupt-parent = <exti> property is not declared in the Device Tree.

Fixes: 0eb9f683336d ("pinctrl: Add IRQ support to STM32 gpios")

Signed-off-by: Fabien Dessenne <fabien.dessenne@foss.st.com>
Change-Id: I9f89e7b24f618e3036c66f549b62018253980997
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/229273
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Alexandre TORGUE <alexandre.torgue@foss.st.com>