dt-bindings: mmc: mmci: Add st,stm32-sdmmc2 compatible
Although this compatible is not used in kernel, as we use the common
MMCI driver, it is used by bootloaders. The U-Boot driver was merged
before the kernel driver and uses this compatible.

To avoid issues when aligning device tree files between kernel and
boot loader, the ST dedicated compatible is added to bindings file.

Signed-off-by: Yann Gautier <yann.gautier@foss.st.com>
Acked-by: Linus Walleij <linus.walleij@linaro.org>
Acked-by: Rob Herring <robh@kernel.org>
Link: https://lore.kernel.org/r/20211210091834.28958-1-yann.gautier@foss.st.com
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
Change-Id: I13817d9f9f5dd522e690c565e163b156d2e1ccde
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/235369
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Christophe KERELLO <christophe.kerello@st.com>