ARM64: dts: ccimx8m: handle both WL_EN/BT_EN lines on Wi-Fi/BT overlays
Enable the reg_rf_bt_en regulator on the Wi-Fi overlay and reg_rf_wl_en on the
BT overlay. This way, if only one of the overlays is applied, but not the
other, the lack of consumer will driver the associated line down (BT_EN or
WL_EN). If not handled, this line may remain tri-stated, causing issues to
the active interface during power-up or resume-from-suspend sequences.

https://onedigi.atlassian.net/browse/DEL-7411

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>