iio: adc: stm32: fix spurious IRQ message and handling
Currently, in case some spurious IRQ is seen in the ADC child driver,
the threaded handler gets called, to display an error message.
We should be returning IRQ_NONE and letting the spurious interrupt
protection kick in, as suggested by Jonathan in [1].

The message also gets removed, as this isn't expected to happen. Another
method could be to move it directly into the isr primary handler before
returning IRQ_NONE.

Another issue this patch addresses is on stm32f4 where "eoc" enable bits
are used as a mask for the "isr" status flags. But these bits on stm32f4
doesn't match together, e.g. eocie is BIT(5), eoc is BIT(1).

[1] https://lore.kernel.org/linux-arm-kernel/20210116175333.4d8684c5@archlinux/

Fixes: 695e2f5c289b ("iio: adc: stm32-adc: fix a regression when using dma and irq")
Suggested-by: Jonathan Cameron <jic23@kernel.org>
Signed-off-by: Fabrice Gasnier <fabrice.gasnier@foss.st.com>
Change-Id: Idbcfe172489f1b0fe410b45890950831aba0747e
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/238199
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Olivier MOYSAN <olivier.moysan@foss.st.com>