mmc: mmci: manage MMC_PM_KEEP_POWER with a flag
Add a disable_keep_power field in variant_data struct. The
MMC_PM_KEEP_POWER flag will be enabled if disable_keep_power is not set.
It is only set to true for stm32_sdmmc variants.

The issue was seen on STM32MP157C-DK2 board, which embeds a wifi chip.
It doesn't correctly support low power on this board. The Wifi chip
awaits an always-on regulator, but it was connected to v3v3 which is off
in low-power sequence. MMC_PM_KEEP_POWER should then be disabled.

The flag can still be enabled through DT property: keep-power-in-suspend.

Signed-off-by: Yann Gautier <yann.gautier@foss.st.com>
Change-Id: I481a88d107246783bad00821b8b95e938519844d
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/239828
Tested-by: Eric FOURMONT <eric.fourmont-ext@st.com>
Tested-by: Christophe ROULLIER <christophe.roullier@st.com>
Reviewed-by: Eric FOURMONT <eric.fourmont-ext@st.com>
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Christophe KERELLO <christophe.kerello@st.com>
Reviewed-by: Christophe ROULLIER <christophe.roullier@st.com>