dmaengine: stm32-dma: fix pause/resume corner issues in double-buffer mode
dmaengine_pause is called by DMA client so, it can be done whatever the
transfer status (just started, ongoing, at the end of a period).
Pausing a DMA channel consists in disabling it, then a Transfer Complete
flag is set.
Transfer Complete flag is also set when a period is completed. Therefore,
in interrupt handler, when TCIF is set, it can be due to a previous pause
or a end of period. The only thing that distinguishes the end of period
from the channel paused, it the SxCR.EN bit. In double-buffer mode,
channel remains enabled when reloading the transfer, while it is disabled
when channel is paused.
This patch also disables Double-Buffer mode in pause instead of resume, so
that in resume DMA is ready to restart the transfer without HW autoreload.

Change-Id: I85c061fef067dedf1a413a1836ccabb440a41c7a
Fixes: dae54988803f ("dmaengine: stm32-dma: add device_pause/device_resume support")
Signed-off-by: Amelie Delaunay <amelie.delaunay@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/239475
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Fabien DESSENNE <fabien.dessenne@foss.st.com>