Revert "caam: Detect if RNG is already initialized."
This fix had a side effect on several units that failed to
initialize SH1 on certain temperature range. To avoid both
the original and the new issues, both RNG State Handles are now
initialized in U-Boot, so there is no need to deinstantiate
them in the Kernel driver during Linux boot process.

This reverts commit 9cd98d99bac13de3493d4d1ef36a0fcf2d6a0a61.

https://onedigi.atlassian.net/browse/DEL-4031
https://onedigi.atlassian.net/browse/CC6UL-1227

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>