drivers: mca: make MCA interface drivers platform-agnostic
When it comes to driver code, the CC6UL and CC8 MCA are identical in all cases,
so there's no need to have separate compatible strings per platform.

In the case of the UART driver, there were some differences at first glance,
specifically in the number of available UARTs - but the driver actually obtains
this number by counting the number of UART child nodes in the device tree, so
it isn't necessary to store this information.

https://onedigi.atlassian.net/browse/DEL-7830

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>