Revert "usb: dwc2: drd: properly handle connector ID when the bus is suspended"
Since "usb: dwc2: drd: fix clock gating on USB role switch"
it's redundant to queue a work that is going to call
dwc2_conn_id_status_change(). It was used to exit the low power mode.
It also caused dwc2_hsotg_core_connect() to be called. It also now cause
the device to connect to the host, even if the gadget driver hasn't been
configured. It results in the host not being able to enumerate:
"device descriptor read/64, error -110".

Signed-off-by: Fabrice Gasnier <fabrice.gasnier@foss.st.com>
Change-Id: I9a3b9b112fcb6da23b5073e2fbaf0a48c6d2889b
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/461203
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>