ARM: dts: stm32: Add a label to sound node for stm32mp157 EV1 and DK2 boards
Because GPIOs are shared by several peripherals, the audio peripherals
can have to be disabled by an overlay. In this case the sound card may
also be disabled to avoid deferred probe that will be never resolved.
To be able to be updated by an overlay the sound node needs a
reference label.

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>
Change-Id: I29a0c357faedfb58d43bb0b7bce5b1695b5219e7
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/246585
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Olivier MOYSAN <olivier.moysan@foss.st.com>
Reviewed-by: Arnaud POULIQUEN <arnaud.pouliquen@st.com>
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>
Tested-by: Arnaud POULIQUEN <arnaud.pouliquen@st.com>