kbuild: support nested composite objects
Kbuild composite object only supports one level of composite objects.  That
is, a composite object may only be composed of real compilable objects. A
composite object may not itself be composed of other composite objects.

Add support for nested composite objects by recursively searching for
composite objects and listing them in $(multi-obj-*) and $(real-obj-*).

This allows to build the qcacld-3.0 external kernel module, that otherwise
fails with an "argument list too long" error due to the deep level of
nested directories and the big number of object files to be linked to
generate the final module.

CC [M]  .../v5.2.0.237G-r0/git/core/pld/src/pld_sdio.o
make[3]: /bin/sh: Argument list too long

The reference for this change is a commit on an old Android kernel for
a Qualcomm device, adapted for kernel 5.15:

https://github.com/LineageOS/android_kernel_oneplus_sm8350/commit/ac9f9a916691495666ab6e990ec061620fdb770d

Signed-off-by: Javier Viguera <javier.viguera@digi.com>