irqchip/stm32-exti: Add multi instance support
The functions in struct syscore_ops don't receive any argument, so
the driver has to keep locally the relevant data.
The global variable 'stm32_host_data' is currently used to keep
the driver data in case of single instance.

Replace the variable with a list in order to add multi instance
support.
Iterate within the list in the syscore_ops functions.
Register the syscore_ops only once per driver.
Remove the useless assignment in stm32_exti_host_init(), as it is
used only by STM32F4 and STM32H7 that do not implement PM.
While there, rework the driver to drop the #if on CONFIG_PM_SLEEP
in favor of using IS_ENABLED(CONFIG_PM_SLEEP).

Signed-off-by: Antonio Borneo <antonio.borneo@foss.st.com>
Change-Id: Id92292f8aa9efc9fef0b448aa3519d7a622a5186
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/257941
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Antonio Maria BORNEO <antonio.borneo@st.com>
Reviewed-by: Alexandre TORGUE <alexandre.torgue@foss.st.com>
Tested-by: Antonio Maria BORNEO <antonio.borneo@st.com>