mfd: mca: complete generalization by removing platform mentions
Substitute CC6UL mentions with KL03, and CC8 mentions with KL17.
Remove platform-dependent Kconfig entries but create a new one
specifically for CC8X MCA. This has the particularity of
the System Controller Unit, that requires to build a special irq handling
driver.
Re-add build support for tamper driver which was missing.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>