iio: adc: stm32: add forced threading support
The current STM32 ADC driver does not support forced threading config.
The ADCs share a common IRQs on F4, H7 SoCs, while each ADC gets its own
IRQ on MP1 SoC. For ADCs sharing the same IRQ, a demux has to be
introduced. This is currently managed through an IRQ domain implementing
a chained IRQ handler in ADC core driver.

In threaded configuration the CPU may stall because the GIC handler may be
called in loop, while the ADC child IRQ handler has never the opportunity
to acknowledge the IRQ. This issue was reported initially in [1].
This problem can be avoided by masking the IRQ in GIC handler.

Add hierarchical IRQ domain to allow management of IRQ parent domain
mask/unmask methods.
Here the IRQ hierarchy is not used to call directly the IRQ child handler
from the parent IRQ handler, as the IRQ demux has still to be performed
for ADCs sharing an IRQ.
The hierarchy is only used to get a reference on parent domain IRQ chip
methods, while the ADC IRQ domain chained handler is kept.

[1] https://lore.kernel.org/linux-arm-kernel/20210112152441.20665-1-a.fatoum@pengutronix.de/
Signed-off-by: Olivier Moysan <olivier.moysan@foss.st.com>
Change-Id: I1a83f80e304a03eff3b5361c2de2f755444c2ed9